Jira Number TM404-75
The changes that were made in the createJob file are the buttons (Schlage, Kwikset, Cancel (the bottom one), and Confirm (the top one)). For the Schlage and Kwikset buttons, if one is selected, the selected button becomes disabled, while the other is enabled. The cancel (the bottom one) makes the Schlage and Kwikset buttons enabled (deselected). If the keyway isn't selected and the user clicks on confirm (the top one), then the app will show an error message.