Jira Numbers TM404-79, TM404-83
TM404-79: After selecting the keyway type, the "Generate" button will be "enabled". After that, the user can press the Confirm button (the button next to "Reset") to see the variables created.
TM404-83: After the "Generate" button is pressed, it calls the generate code. When the keyway type is selected and the number of change keys is typed, the confirm button (the button next to "Reset") will be "enabled". After pressing Confirm (the button next to "Reset"), the variables will appear on the page.