Jira Numbers TM404-102, TM404-105, TM404-104
TM404-102: If the master key level is 2, then additional textfields and buttons will appear.
TM404-105: Depending on the master key level, certain information will appear. Some information will not appear because it is not part of master key levels 1 or 2. The information is color coded for identifying certain information.
TM404-104: If the user enters in anything that isn't a number, a negative number, zero, or the number is not entered properly, an error message appears.