Jira Number TM404-110
One minor change that was made in the createJob code was the second background on lines 551-553 being deleted because when I run my code with the second background, the textfield for number of change keys and changing the master key level will not work. Although this may change the appearance of the page, it will not change the functionality of the page.