Jira Number TM404-112
Changes that were made in the create job page are adding a scrollview feature and moving components 7, 8, and 9 outside the scrollview because the code before would cover up the "Num Keys for SMKeys" textfields when the simulator's keyboard would appear. After implementing the scrollview, the components 7, 8, and 9 were moved outside the scrollview because component 9 (the "Cancel" button) would block the displayed keys combinations. Other changes includes deleting the two "Reset" buttons, disabling the "Add Combo" button when the textfields appear, renaming the "Cancel" button to "Clear" and moving the "Clear" button next to the "Confirm" button.